Amendments on copyright for 0003: Guidelines on Licensing and DCO
Changes the wording of copyright notices to read "The <project>
Contributors" in LICENSE files, and adds a `rustfmt` header check
template which doesn't require a fixed copyright holder.

Reasoning: in the jurisdictions we're operating in, copyright is not
transferrable. While contributions by employees fall under the copyright
of the employer, this is not true for external contributions: if the
work is copyrightable (ie. "substantial"), they can add their own name
to the license header of the file(s) they created / modified.

Signed-off-by: Kim Altintop <kim@monadic.xyz>